Move log related commands to `logs` namespace.
