+++ lib/api/client-commands/logs/captureBrowserConsoleLogs.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('./_base-command.js');
-const {Logger} = require('../../utils');
+const ClientCommand = require('../_base-command.js');
+const {Logger} = require('../../../utils');
 
 /**
  * Listen to the `console` events (ex. `console.log` event) and register callback to process the same.
@@ -26,9 +26,11 @@ const {Logger} = require('../../utils');
  * @moreinfo nightwatchjs.org/guide/running-tests/capture-console-messages.html
  */
 class StartCapturingLogs extends ClientCommand {
+  static get namespacedAliases() {
+    return 'captureBrowserConsoleLogs';
+  }
 
   performAction(callback) {
-
     if (!this.api.isChrome()  && !this.api.isEdge()) {
       const error = new Error('The command .captureBrowserConsoleLogs() is only supported in Chrome and Edge drivers');
       Logger.error(error);
