+++ lib/api/client-commands/logs/captureBrowserExceptions.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('./_base-command.js');
-const {Logger} = require('../../utils');
+const ClientCommand = require('../_base-command.js');
+const {Logger} = require('../../../utils');
 
 /**
  * Catch the JavaScript exceptions thrown in the browser.
@@ -30,9 +30,11 @@ const {Logger} = require('../../utils');
  * @moreinfo nightwatchjs.org/guide/running-tests/catch-js-exceptions.html
  */
 class CatchJsExceptions extends ClientCommand {
+  static get namespacedAliases() {
+    return 'captureBrowserExceptions';
+  }
 
   performAction(callback) {
-
     if (!this.api.isChrome() && !this.api.isEdge()) {
       const error = new Error('The command .captureBrowserExceptions() is only supported in Chrome and Edge drivers');
       Logger.error(error);
