+++ lib/api/client-commands/logs/getLog.js
@@ -1,4 +1,4 @@
-const ClientCommand = require('./_base-command.js');
+const ClientCommand = require('../_base-command.js');
 
 /**
  * Gets a log from Selenium.
@@ -22,6 +22,10 @@ const ClientCommand = require('./_base-command.js');
  * @see getLogTypes
  */
 class GetLog extends ClientCommand {
+  static get namespacedAliases() {
+    return 'getLog';
+  }
+
   get returnsFullResultObject() {
     return false;
   }
@@ -30,8 +34,8 @@ class GetLog extends ClientCommand {
     return false;
   }
 
-  performAction(actionCallback) {
-    this.api.sessionLog(this.typeString, actionCallback);
+  performAction(callback) {
+    this.transportActions.getLogContents(this.typeString, callback);
   }
 
   command(typeString = 'browser', callback) {
