+++ lib/api/client-commands/logs/getLogTypes.js
@@ -1,4 +1,4 @@
-const ClientCommand = require('./_base-command.js');
+const ClientCommand = require('../_base-command.js');
 
 /**
  * Gets the available log types. More info about log types in WebDriver can be found here: https://github.com/SeleniumHQ/selenium/wiki/Logging
@@ -19,6 +19,10 @@ const ClientCommand = require('./_base-command.js');
  * @see sessionLogTypes
  */
 class GetLogTypes extends ClientCommand {
+  static get namespacedAliases() {
+    return 'getLogTypes';
+  }
+
   get returnsFullResultObject() {
     return false;
   }
@@ -27,10 +31,9 @@ class GetLogTypes extends ClientCommand {
     return false;
   }
 
-  performAction(actionCallback) {
-    this.api.sessionLogTypes(actionCallback);
+  performAction(callback) {
+    this.transportActions.getSessionLogTypes(callback);
   }
-
 }
 
 module.exports = GetLogTypes;
