+++ lib/api/client-commands/logs/isAvailable.js
@@ -1,4 +1,4 @@
-const ClientCommand = require('./_base-command.js');
+const ClientCommand = require('../_base-command.js');
 
 /**
  * Utility command to test if the log type is available.
@@ -19,10 +19,14 @@ const ClientCommand = require('./_base-command.js');
  * @see getLogTypes
  */
 class IsLogAvailable extends ClientCommand {
-  performAction(actionCallback) {
+  static get namespacedAliases() {
+    return 'isLogAvailable';
+  }
+
+  performAction(callback) {
     const {typeString} = this;
 
-    this.api.getLogTypes(function(types) {
+    this.transportActions.getSessionLogTypes(function(types) {
       let isAvailable;
 
       try {
@@ -31,7 +35,7 @@ class IsLogAvailable extends ClientCommand {
         isAvailable = false;
       }
 
-      actionCallback.call(this, isAvailable);
+      callback.call(this, isAvailable);
     });
   }
